requirements/req.txt
beautifulsoup4==4.8.0
bs4==0.0.1
Django==2.2.5
django-background-tasks==1.2.0
django-cache-memoize==0.1.6
django-compat==1.0.15
django-debug-toolbar==2.0
et-xmlfile==1.0.1
jdcal==1.4.1
lxml==4.4.1
numpy==1.17.2
openpyxl==3.0.0
pandas==0.25.1
Pillow==6.1.0
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
